:recycle: PairRoomOnboarding 페이지 구조 개선
