--- frontend/src/components/PairRoomOnBoarding/OnboardingModal/OnboardingModal.styles.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const Layout = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  flex: 1;
-`;
