--- frontend/src/components/PairRoomOnBoarding/OnboardingModal/constants.ts
@@ -1,5 +1,0 @@
-export const TIMER_OPTIONS = [
-  { label: '10분', value: '10' },
-  { label: '15분', value: '15' },
-  { label: '30분', value: '30' },
-];
