+++ frontend/src/components/PairRoomOnBoarding/ProgressBar/ProgressBar.styles.ts
