+++ frontend/src/components/PairRoomOnBoarding/TimerSettingSection/TimerSettingSection.tsx
@@ -1,22 +1,26 @@
 import { useState } from 'react';
 
-import { RiInformation2Line } from 'react-icons/ri';
-
 import Button from '@/components/common/Button/Button';
 import Input from '@/components/common/Input/Input';
 import { Modal } from '@/components/common/Modal';
-import { TIMER_OPTIONS } from '@/components/PairRoomOnBoarding/OnboardingModal/constants';
+import InformationBox from '@/components/PairRoomOnBoarding/InformationBox/InformationBox';
 
 import { validateTime } from '@/utils/PairRoomOnboarding/validate';
 
-import * as S from './Steps.styles';
+import * as S from './TimerSettingSection.styles';
 
-interface TimerSettingProps {
+const OPTIONS = [
+  { label: '10분', value: '10' },
+  { label: '15분', value: '15' },
+  { label: '30분', value: '30' },
+];
+
+interface TimerSettingSectionProps {
   timer: string;
   onTimer: (time: string) => void;
 }
 
-const TimerSetting = ({ timer, onTimer }: TimerSettingProps) => {
+const TimerSettingSection = ({ timer, onTimer }: TimerSettingSectionProps) => {
   const [isSelf, setIsSelf] = useState(false);
 
   const handleTimer = (option: string) => {
@@ -32,17 +36,13 @@ const TimerSetting = ({ timer, onTimer }: TimerSettingProps) => {
   return (
     <S.Layout>
       <Modal.Header title="타이머 설정" subTitle="타이머 시간을 설정해 주세요" />
-      <S.InformationWrapper>
-        <S.InformationTitle>
-          <RiInformation2Line size="2rem" />왜 타이머 시간을 설정해야 하나요?
-        </S.InformationTitle>
-        <S.InformationDescription>
-          정기적인 역할 교대는 피드백을 주고받을 수 있는 자연스러운 기회를 제공합니다. 이는 코드 품질을 높이고, 문제를
-          조기에 발견하여 수정할 수 있게 합니다.
-        </S.InformationDescription>
-      </S.InformationWrapper>
-      <S.SettingsContainer>
-        {TIMER_OPTIONS.map((option) => (
+      <InformationBox
+        title="왜 타이머 시간을 설정해야 하나요?"
+        description="정기적인 역할 교대는 피드백을 주고받을 수 있는 자연스러운 기회를 제공합니다. 이는 코드 품질을 높이고, 문제를
+          조기에 발견하여 수정할 수 있게 합니다."
+      />
+      <S.ButtonContainer>
+        {OPTIONS.map((option) => (
           <Button
             key={option.value}
             color="primary"
@@ -53,7 +53,7 @@ const TimerSetting = ({ timer, onTimer }: TimerSettingProps) => {
             {option.label}
           </Button>
         ))}
-        <S.TimeInputWrapper>
+        <S.InputContainer>
           <Button key="직접 설정" color="primary" size="md" filled={isSelf} onClick={handleIsSelf}>
             직접 설정
           </Button>
@@ -63,16 +63,16 @@ const TimerSetting = ({ timer, onTimer }: TimerSettingProps) => {
               $css={S.inputStyles}
               value={timer}
               placeholder="타이머 시간 (분)"
-              status={!validateTime(Number(timer)) ? 'error' : 'default'}
-              message={!validateTime(Number(timer)) ? '0 이상의 숫자를 입력해 주세요.' : ''}
+              status={!validateTime(timer) ? 'error' : 'default'}
+              message={!validateTime(timer) ? '0 이상의 숫자를 입력해 주세요.' : ''}
               disabled={!isSelf}
               onChange={(event) => onTimer(event.target.value)}
             />
           )}
-        </S.TimeInputWrapper>
-      </S.SettingsContainer>
+        </S.InputContainer>
+      </S.ButtonContainer>
     </S.Layout>
   );
 };
 
-export default TimerSetting;
+export default TimerSettingSection;
