+++ frontend/src/pages/PairRoomOnboarding/PairRoomOnboarding.type.ts
