+++ frontend/src/utils/PairRoomOnboarding/validate.ts
@@ -1,6 +1,7 @@
-export const validateTime = (time: number) => {
-  if (Number.isNaN(time)) return false;
-  if (time < 0) return false;
+export const validateTime = (time: string) => {
+  if (time === '') return false;
+  if (Number.isNaN(Number(time))) return false;
+  if (Number(time) < 0) return false;
 
   return true;
 };
